+++ HISTORY.rst
@@ -21,6 +21,7 @@ API changes:
 * Template class replaced by a Renderer class. [cjerdonek]
 * ``Loader.load_template()`` changed to ``Loader.get()``. [cjerdonek]
 * Removed output_encoding options. [cjerdonek]
+* Removed automatic use of markupsafe, if available. [cjerdonek]
 
 Bug fixes: